Add support for exporting multiple clients and mark the old client as deprecated
